+++ packages/docs/client/components/Props/index.js
@@ -1,14 +1,65 @@
 import React, { useMemo, useState } from 'react'
 import { ScrollView } from 'react-native'
 import { observer, $root, useComponentId } from 'startupjs'
-import './index.styl'
 import { themed, Button, Row, Div } from '@startupjs/ui'
+import parsePropTypes from 'parse-prop-types'
 import Constructor from './Constructor'
 import Renderer from './Renderer'
+import './index.styl'
+
+function parseEntries (entries) {
+  return entries.map(entry => {
+    let meta = entry[1]
+    return {
+      name: entry[0],
+      type: meta.type.name,
+      defaultValue: meta.defaultValue && meta.defaultValue.value,
+      possibleValues: meta.type.value,
+      isRequired: meta.required
+    }
+  })
+}
+
+function useEntries ({ Component, props = {}, extraParams }) {
+  return useMemo(() => {
+    const entries = parseEntries(Object.entries(parsePropTypes(Component)))
+      .filter(entry => entry.name[0] !== '_') // skip private properties
+      .map(item => {
+        if (props[item.name] !== undefined) {
+          item.value = props[item.name] // add property value to Renderer
+        }
+        if (extraParams?.[item.name]) {
+          item.extraParams = extraParams?.[item.name]
+        }
+        return item
+      })
+    return entries
+  }, [])
+}
+
+async function useInitDefaultProps ({ entries, $theProps }) {
+  if ($theProps.get()) return
+  $theProps.set('', {})
+
+  const promises = []
+
+  for (const prop of entries) {
+    if (prop.defaultValue !== undefined || prop.value !== undefined) {
+      // NOTE: Due to a racer patch, last argument cannot be a function
+      // because it will be used as a callback of `$props.set`,
+      // so we use null to avoid this behavior when defaultValue is function
+      promises.push($theProps.set(prop.name, prop.value || prop.defaultValue, null))
+    }
+  }
+
+  await Promise.all(promises)
+}
 
 export default observer(themed(function PComponent ({
   Component,
   $props,
+  props,
+  extraParams,
   componentName,
   showGrid,
   style,
@@ -19,6 +70,7 @@ export default observer(themed(function PComponent ({
 }) {
   const [block, setBlock] = useState(!!defaultBlock)
   const componentId = useComponentId()
+
   const $theProps = useMemo(() => {
     if (!$props) {
       return $root.scope(`_session.Props.${componentId}`)
@@ -26,15 +78,21 @@ export default observer(themed(function PComponent ({
       return $props
     }
   }, [$props])
-  $theProps.setNull('', {})
+
+  const entries = useEntries({ Component, props, extraParams })
+  useInitDefaultProps({ entries, $theProps })
 
   return pug`
     Div.root(style=style)
       ScrollView.top(styleName=[theme])
-        Constructor(Component=Component $props=$theProps)
-      ScrollView.bottom(
-        styleName=[theme, { showSizes }]
-      )
+        Constructor(
+          Component=Component
+          $props=$theProps
+          entries=entries
+          props=props
+        )
+
+      ScrollView.bottom(styleName=[theme, { showSizes }])
         Renderer(
           Component=Component
           props=$theProps.get()