+++ packages/docs/client/components/Sandbox/index.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { observer, $root } from 'startupjs'
-import Props from '../Props'
 import { H4 } from '@startupjs/ui'
+import Props from '../Props'
 import {
   useShowGrid,
   useShowSizes,
@@ -12,7 +12,7 @@ import './index.styl'
 
 const MODELS = new WeakMap()
 
-export default observer(function Sandbox ({ Component, block }) {
+export default observer(function Sandbox ({ Component, props, extraParams, block }) {
   const [showGrid] = useShowGrid()
   const [showSizes] = useShowSizes()
   const [validateWidth] = useValidateWidth()
@@ -32,6 +32,8 @@ export default observer(function Sandbox ({ Component, block }) {
       showGrid=showGrid
       validateWidth=validateWidth
       $props=getUniqModel(Component)
+      props=props
+      extraParams=extraParams
       block=block
     )
   `