+++ packages/ui/components/SmartSidebar/index.js
@@ -41,8 +41,14 @@ function SmartSidebar ({
   let open
   let onChange
   ;({ open, onChange } = useBind({ $open: $open, open, onChange }))
+  const [, $localOpen] = useValue()
 
   let [fixedLayout, $fixedLayout] = useValue(isFixedLayout(fixedLayoutBreakpoint))
+
+  useLayoutEffect(() => {
+    const newOpen = disabled ? false : open
+    $localOpen.setDiff(newOpen)
+  }, [open, disabled])
 
   useLayoutEffect(() => {
     if (disabled) {
@@ -83,7 +89,7 @@ function SmartSidebar ({
       Sidebar(
         style=style
         sidebarStyle=sidebarStyle
-        $open=$open
+        $open=$localOpen
         position=position
         width=width
         renderContent=renderContent
@@ -91,7 +97,7 @@ function SmartSidebar ({
     else
       DrawerSidebar(
         style=style
-        $open=$open
+        $open=$localOpen
         position=position
         width=width
         renderContent=renderContent
@@ -103,7 +109,7 @@ function SmartSidebar ({
 
 SmartSidebar.defaultProps = {
   defaultOpen: false,
-  disalbed: false,
+  disabled: false,
   fixedLayoutBreakpoint: FIXED_LAYOUT_BREAKPOINT,
   position: 'left',
   width: 264
@@ -114,7 +120,7 @@ SmartSidebar.propTypes = {
   children: PropTypes.node,
   $open: PropTypes.object,
   defaultOpen: PropTypes.bool,
-  disalbed: PropTypes.bool,
+  disabled: PropTypes.bool,
   fixedLayoutBreakpoint: PropTypes.number,
   position: PropTypes.oneOf(['left', 'right']),
   width: PropTypes.number,