+++ packages/ui/components/popups/Popover/Deprecated/Arrow/index.js
@@ -7,11 +7,15 @@ export default function Arrow ({
   geometry,
   validPosition
 }) {
+  if (geometry && geometry.validPlacement) {
+    [validPosition] = geometry.validPlacement.split('-')
+  }
+
   const _arrowStyle = StyleSheet.flatten([
     style,
     {
-      left: geometry.arrowLeftPosition || 0,
-      top: geometry.arrowTopPosition || 0
+      left: geometry ? geometry.arrowLeftPosition : 0,
+      top: geometry ? geometry.arrowTopPosition : 0
     }
   ])