+++ packages/ui/components/popups/Popover/animate.js
@@ -7,7 +7,7 @@ export default {
     durationOpen,
     animateType,
     animateStates
-  }, callback) {
+  }) {
     const validPlacement = geometry.validPlacement
     const [position, attachment] = validPlacement.split('-')
 
@@ -35,7 +35,7 @@ export default {
           duration: durationOpen,
           useNativeDriver: false
         })
-      ]).start(callback)
+      ])
     }
 
     if (animateType === 'scale') {
@@ -85,7 +85,7 @@ export default {
           duration: durationOpen,
           useNativeDriver: false
         })
-      ]).start(callback)
+      ])
     }
   },
 
@@ -95,7 +95,15 @@ export default {
     durationClose,
     animateType,
     animateStates
-  }, callback) {
+  }) {
+    if (!geometry) {
+      return Animated.timing(animateStates.opacity, {
+        toValue: 0,
+        duration: durationClose,
+        useNativeDriver: false
+      })
+    }
+
     animateStates.height.setValue(contentInfo.height)
     const [position, attachment] = (geometry.validPlacement || '').split('-')
 
@@ -123,7 +131,7 @@ export default {
           duration: durationClose,
           useNativeDriver: false
         })
-      ]).start(callback)
+      ])
     }
 
     if (animateType === 'scale') {
@@ -182,7 +190,7 @@ export default {
           duration: durationClose,
           useNativeDriver: false
         })
-      ]).start(callback)
+      ])
     }
   }
 }