+++ packages/ui/components/typography/Span/index.js
@@ -56,7 +56,7 @@ styl`
 
   // ----- COMPONENT
 
-  _variants = ('default' 'h1' 'h2' 'h3' 'h4' 'h5' 'h6' 'description') // H1-H6 DEPRECATED
+  _variants = ('default' 'h1' 'h2' 'h3' 'h4' 'h5' 'h6' 'description') // H1-H6, description is DEPRECATED
 
   .root
     font()